+++ Shared/Models/DynamicIntent/Person.swift
@@ -21,19 +21,55 @@
 // SOFTWARE.
 
 import Foundation
-import SwiftData
 
-@Model
-final class Product {
-    @Attribute(.unique) var id = UUID().uuidString
+struct Person: Codable, Hashable, Identifiable {
+    var id = UUID().uuidString
     var name: String
-    var creationDate: Date
+    var dateOfBirth: Date
+}
 
-    init(
-        name: String = "Product \(Int.random(in: 1 ... 99))",
-        creationDate: Date = .now
-    ) {
-        self.name = name
-        self.creationDate = creationDate
+// MARK: - Helpers
+
+extension Person {
+    static func getAll() -> [Self] {
+        let key = UserDefaultKey.persons
+        guard let persons: [Self] = UserDefaults.appGroup.getArray(forKey: key) else {
+            let persons: [Self] = .defaultFriends
+            UserDefaults.appGroup.setArray(persons, forKey: key)
+            return persons
+        }
+        return persons
     }
+}
+
+// MARK: - Convenience
+
+extension Person {
+    init?(identifier: String) {
+        if let person = Self.getAll().first(where: { $0.id == identifier }) {
+            self = person
+        } else {
+            return nil
+        }
+    }
+}
+
+// MARK: - Data
+
+extension [Person] {
+    static let defaultFriends: Self = [
+        .friend1, .friend2
+    ]
+}
+
+extension Person {
+    static let friend1: Self = .init(
+        name: "Friend 1",
+        dateOfBirth: .now.adding(.month, value: -2)
+    )
+
+    static let friend2: Self = .init(
+        name: "Friend 2",
+        dateOfBirth: .now.adding(.year, value: -3)
+    )
 }
