+++ Shared/Models/LiveActivity/DeliveryAttributes.swift
